Fix PNG not working in release mode
It was caused by some optimisation that wasn't done in debug mode (which made me not notice a bad practice i used) but that was present in release mode. It was harder to debug as in release mode there is much less debug information (obviously) even in release-safe mode.

The problem actually was that i allocated a ChunkStream on the stack and just took the reader out of it, however the reader depended on the ChunkStream to not be overwritten. And for some reason it was ok in Debug mode but some optimization in release mode caused that problem to be visible. It might also be caused by the fact that by storing the reader directly, i thought it was ok to store chunks as constants when i use Chunk.toStruct() :neutral_face: (this doesn't affect reading IDAT as i use the data field directly) and so release mode probably also had some optimization for constant variables too.

In total this was just caused by bad practice that went unnoticed in Debug mode. (and to the Zig compiler folks, that wouldn't be easy to fix unless doing like when Debug writes 0xaa to undefined values, but adapt it to find those sort of bugs)